requirements.txt
# Python dependencies for MediLens Fine-tuning
google-generativeai>=0.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
pathlib
logging
